Update example.js
I noticed that the example didn't have a setdoorstate for closing the door. I tried using state 2 to close the door but that threw an error. After inspecting what the myliftmaster.com client side post does, I used the state of 0 to close the door which worked. 
Please verify if you still have a myQ device and have interest in the project. Thank-you for posting this repository. It's been useful.